LayoutTests/storage/indexeddb/intversion-and-setversion-expected.txt

Test that int version upgrades a string version and that string versions are then disallowed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "intversion-and-setversion.html"
indexedDB.deleteDatabase(dbname)
FAIL String(request) should be [object IDBOpenDBRequest]. Was [object IDBVersionChangeRequest].

Call open with no version parameter:
indexedDB.open(dbname)

openSuccess():
db = event.target.result
PASS db.version is ""
request = db.setVersion("some version")

inSetVersion():
PASS db.version is "some version"

setVersionDone():
PASS db.version is "some version"
db.close()
indexedDB.open(dbname)
request.onsuccess = secondOpen

secondOpen():
db = event.target.result
PASS String(db) is "[object IDBDatabase]"
PASS db.version is "some version"
db.close()
request = indexedDB.open(dbname, 2)

firstOpenWithVersion():
FAIL didTransactionComplete should be true. Was false.
db = event.target.result
PASS String(db) is "[object IDBDatabase]"
PASS request.transaction is null
FAIL db.version should be 2 (of type number). Was some version (of type string).
FAIL String(request) should be [object IDBOpenDBRequest]. Was [object IDBRequest].
request = db.setVersion("string version 2")
PASS String(request) is "[object IDBVersionChangeRequest]"
request.onerror = setVersionNotAllowed
FAIL Success function called unexpectedly.
PASS successfullyParsed is true

TEST COMPLETE
